Revise interface significantly.
Revise public interface so that only folders have a constructor -- all
other objects must be created as part of an existing folder.
This means that paths need no longer be optional components -- the
object is never constructed if there is no valid path.
Add some convenience methods for converting to snake/kebab case.